+++ jobbergate-agent/jobbergate_agent/settings.py
@@ -39,7 +39,6 @@ class Settings(BaseSettings):
 
     # OIDC config for machine-to-machine security
     OIDC_DOMAIN: str
-    OIDC_AUDIENCE: str = "https://apis.omnivector.solutions"
     OIDC_CLIENT_ID: str
     OIDC_CLIENT_SECRET: str
     OIDC_USE_HTTPS: bool = True