+++ jobbergate-agent/CHANGELOG.md
@@ -5,6 +5,7 @@ This file keeps track of all notable changes to jobbergate-agent
 ## Unreleased
 - Changed auto-update task to reuse current scheduler instead of creating a new one
 - Fixed environment variables from the machine running the agent propagating to slurm jobs (notice `--export=ALL` is the default behavior for sbatch)
+- Removed the `OIDC_AUDIENCE` setting
 
 ## 5.3.0 -- 2024-09-09